Put types under a common package
