+++ verify/types.go
@@ -8,8 +8,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/digitorus/timestamp"
-	"golang.org/x/crypto/ocsp"
+	"github.com/digitorus/pdfsign/common"
 )
 
 // VerifyOptions contains options for PDF signature verification
@@ -54,75 +53,26 @@ type VerifyOptions struct {
 	HTTPTimeout time.Duration
 }
 
-// SignatureInfo contains information about the signer and signature
-// (not related to validation)
-type SignatureInfo struct {
-	Name          string               `json:"name"`
-	Reason        string               `json:"reason"`
-	Location      string               `json:"location"`
-	ContactInfo   string               `json:"contact_info"`
-	SignatureTime *time.Time           `json:"signature_time,omitempty"`
-	TimeStamp     *timestamp.Timestamp `json:"time_stamp"`
-	DocumentHash  string               `json:"document_hash"`
-	SignatureHash string               `json:"signature_hash"`
-	HashAlgorithm string               `json:"hash_algorithm"`
-}
-
 // SignatureValidation contains validation results and technical details
 // (not about the signer's intent)
 type SignatureValidation struct {
-	ValidSignature     bool          `json:"valid_signature"`
-	TrustedIssuer      bool          `json:"trusted_issuer"`
-	RevokedCertificate bool          `json:"revoked_certificate"`
-	Certificates       []Certificate `json:"certificates"`
-	TimestampStatus    string        `json:"timestamp_status,omitempty"`
-	TimestampTrusted   bool          `json:"timestamp_trusted"`
-	VerificationTime   *time.Time    `json:"verification_time"`
-	TimeSource         string        `json:"time_source"`
-	TimeWarnings       []string      `json:"time_warnings,omitempty"`
+	ValidSignature     bool                 `json:"valid_signature"`
+	TrustedIssuer      bool                 `json:"trusted_issuer"`
+	RevokedCertificate bool                 `json:"revoked_certificate"`
+	Certificates       []common.Certificate `json:"certificates"`
+	TimestampStatus    string               `json:"timestamp_status,omitempty"`
+	TimestampTrusted   bool                 `json:"timestamp_trusted"`
+	VerificationTime   *time.Time           `json:"verification_time"`
+	TimeSource         string               `json:"time_source"`
+	TimeWarnings       []string             `json:"time_warnings,omitempty"`
 }
 
 type Response struct {
 	Error string
 
-	DocumentInfo DocumentInfo
+	DocumentInfo common.DocumentInfo
 	Signatures   []struct {
-		Info       SignatureInfo       `json:"info"`
-		Validation SignatureValidation `json:"validation"`
+		Info       common.SignatureInfo `json:"info"`
+		Validation SignatureValidation  `json:"validation"`
 	}
-}
-
-type Certificate struct {
-	Certificate          *x509.Certificate `json:"certificate"`
-	VerifyError          string            `json:"verify_error"`
-	KeyUsageValid        bool              `json:"key_usage_valid"`
-	KeyUsageError        string            `json:"key_usage_error,omitempty"`
-	ExtKeyUsageValid     bool              `json:"ext_key_usage_valid"`
-	ExtKeyUsageError     string            `json:"ext_key_usage_error,omitempty"`
-	OCSPResponse         *ocsp.Response    `json:"ocsp_response"`
-	OCSPEmbedded         bool              `json:"ocsp_embedded"`
-	OCSPExternal         bool              `json:"ocsp_external"`
-	CRLRevoked           time.Time         `json:"crl_revoked"`
-	CRLEmbedded          bool              `json:"crl_embedded"`
-	CRLExternal          bool              `json:"crl_external"`
-	RevocationWarning    string            `json:"revocation_warning,omitempty"`
-	RevocationTime       *time.Time        `json:"revocation_time,omitempty"` // When the certificate was revoked (if applicable)
-	RevokedBeforeSigning bool              `json:"revoked_before_signing"`    // Whether revocation occurred before signing
-}
-
-// DocumentInfo contains document information.
-type DocumentInfo struct {
-	Author     string `json:"author"`
-	Creator    string `json:"creator"`
-	Hash       string `json:"hash"`
-	Name       string `json:"name"`
-	Permission string `json:"permission"`
-	Producer   string `json:"producer"`
-	Subject    string `json:"subject"`
-	Title      string `json:"title"`
-
-	Pages        int       `json:"pages"`
-	Keywords     []string  `json:"keywords"`
-	ModDate      time.Time `json:"mod_date"`
-	CreationDate time.Time `json:"creation_date"`
 }
